added combat screen battlefield; moved combat data to monad

+++ common/buttons.go
@@ -10,8 +10,8 @@ import (
 )
 
 type Button struct {
-	Width, Height int
-	X, Y          int
+	Width, Height float32
+	X, Y          float32
 	Text          string
 	Active        bool
 }
@@ -27,11 +27,11 @@ func (m *Button) Draw(screen *ebiten.Image) {
 	vector.DrawFilledRect(screen, float32(m.X), float32(m.Y), float32(m.Width), float32(m.Height), buttonColor, false)
 
 	bounds := font.MeasureString(MenuFont, m.Text)
-	textX := m.X + (m.Width-bounds.Floor())/2
-	text.Draw(screen, m.Text, MenuFont, textX, m.Y+m.Height/2+5, color.White)
+	textX := m.X + (m.Width-float32(bounds.Floor()))/2
+	text.Draw(screen, m.Text, MenuFont, int(textX), int(m.Y+m.Height/2+5), color.White)
 }
 
-func (m *Button) GetBounds() (int, int, int, int) {
+func (m *Button) GetBounds() (float32, float32, float32, float32) {
 	return m.X, m.Y, m.X + m.Width, m.Y + m.Height
 }
